+++ package.json
@@ -13,6 +13,9 @@
     "dev": "nodemon server.js"
   },
   "dependencies": {
+    "@types/express": "^4.17.2",
+    "@types/graphql-iso-date": "^3.3.3",
+    "@types/lodash": "^4.14.149",
     "apollo-server-express": "^1.4.0",
     "body-parser": "^1.19.0",
     "express": "^4.17.1",
@@ -23,7 +26,8 @@
     "nodemon": "^2.0.2",
     "pg": "^7.17.1",
     "pg-hstore": "^2.3.3",
-    "sequelize": "^5.21.3"
+    "sequelize": "^5.21.3",
+    "typescript": "^3.7.5"
   },
   "homepage": "https://github.com/uwblueprint/paramedics-web",
   "devDependencies": {
